Intel/gfortran support + static library + missing files + refactor + CI (Windows, macOS, Linux)
The original NREL EnergyPlusRelease repository did not build out of the box.

This version of EnergyPlus (EnergyPlusFortran) builds 'out of the box' and adds support for Windows, macOS, and Linux for both GCC and Intel compilers. Additionally to the standard energyplus executable, a static library, containing all the public functions and subroutines, is created at compile time. Build instructions for compiling under Windows, macOS, and Linux with CMake were added to the README file.

Missing files and dependencies required to compile the executables were added to the repository;
these files were taken from a fork of NREL's EnergyPlus Release (nrgsim) and from SQLite (sqlite3.c was missing from the original repository so SQLite amalgamation version 3.18.0 was added).

Continuous Integration (CI) through AppVeyor for Intel compilers under Windows, and Travis CI for GCC gfortran and Intel compilers under macOS and Linux, was added.

The folder structure was also changed so that files and folders could be more easily located.

A few functions were made public so that they could be used externally (NumCurves and PerfCurve in CurveManager.f90 and Schedule in ScheduleManager.f90).

DElight and ExternalInterface were not included with the CMake refactoring and are still not working. Lib files were removed as they were version specific and should be compiled at compile time.

A few corrections were made to a small number of Fortran files to make them compatible with different versions of gfortran.