Encode to utf-8 + add opt to disable cache + update Intel lic (#2)
All files which were encoded as iso-8859-1 or similar were re encoded as utf-8. An option to disable the cache in the psychrometric routines was added to the CMakeLists.txt in src and this option can be now passed during configure. The license number for the Intel compiler for Windows was updated.