add reference doc generation with ford
The input file to generate documentation from the code base using ford was created and added to the docs folder. The link to the reference documentation was added in README.md. The e+ logo and the e+ favicon were added into the media folder under docs to be used by ford